[PATCH] 3c59x: bring back the `enable_wol' option
Some machines appear to have BIOS problems which are causing 3c59x adapters
to come up in a powered-off state when WOL and PM are enabled.

So bring back the 2.4 `enable_wol' module option which disables wake-on-lan
unless the user specifically asked for it.

BKrev: 403441e22GWRxJlU2PfM7ZByivO5nw