[PATCH] fixes to ColdFire/5407 startup code
Correct the cache setup bits for the 5407. This enables the write
buffers properly (despite what the previous comment said). This
combined with fixed cache flushing code provides a nice performance
boost on the 5407.

Also fix the ROMfs setup to only move the ROMfs region if it is actually
configured.

BKrev: 4034546eyIF04NaBdFf9E8T5Pgw4gQ