[PATCH] Fix a DMA underrun problem with pmac IDE
This fixes the behaviour of the pmac "macio" IDE driver when a DMA
transfer happen to get less data out of the device than expected when
setting up the DMA commands (the device underruns).  This is very common
with recent ATAPI stuffs and used to cause problem & disable DMA.  This
patch fixes the way we handle that condition, thus also fixing DVD
burning on a bunch of recent pmacs.

BKrev: 40359cf06jpocUH0LFLVMh3J2fgaJA