Merge bk://gkernel.bkbits.net/net-drivers-2.5 into ppc970.osdl.org:/home/torvalds/v2.5/linux
2004/02/19 22:56:14-05:00 jgarzik
Merge redhat.com:/spare/repo/netdev-2.6/e100-2
into redhat.com:/spare/repo/net-drivers-2.5

2004/02/06 23:38:05-05:00 jgarzik
Merge redhat.com:/spare/repo/linux-2.5
into redhat.com:/spare/repo/netdev-2.6/e100-2

2004/01/19 18:11:34-05:00 jgarzik
Merge redhat.com:/spare/repo/netdev-2.6/e100-1
into redhat.com:/spare/repo/netdev-2.6/e100-2

2004/01/13 15:30:03-05:00 scott.feldman
[netdrvr e100] copyright + trailing blanks + misc

* Misc: 2004 copyright, remove trailing white space, remove some
  unused symbols.

2004/01/13 15:29:55-05:00 scott.feldman
[netdrvr e100] fix slab corruption

* Addresses two problems, both resulting in slab corruption: 1)
  driver indicating skb while HW is still DMA'ing (ouch!), 2)
  driver not stopping receiver activity before downing i/f.
  Fix is 1) wait for RNR (receiver-no-resources) interrupt
  before restarting receiver, 2) reseting HW to stop receiver
  before stopping i/f.

This issue was also reproducible with eepro100.  You need to turn off
the copybreak, and reduce the number of descriptors to 4.  Then bang on
it with pktgen with 60-byte packets, with slab debugging enabled.

For e100-3.0.x, the issue was a lot easier to reproduce with NAPI,
because NAPI polls independently of where the HW is at, so it's easier
for us to catch HW in the middle of finishing off the last Rx (as
it runs out of resources) and asking HW if it's idle.  Checking the
RU status is not-reliable!  That's the problem, and the mistake both
eepro100 and e100-3.0.x were making.

The solution is rely on RNR interrupts as the only indicator that HW is
truly done, and then we're ready to restart the RU.  We should only get
RNR interrupts when we overrun the Rx ring.  With NAPI, if the ring is
overrun, we'll post RNR, but not restart the RU until we're out of
polling.  Without NAPI, we'll restart the RU as soon as we get RNR.

I ran some 24-hour tests with and without NAPI (with 4 descriptors)
and didn't get any corruption.  Prior to this patch, I would get many
errors about slab corruption.

Also, the patch is larger than you might expect, but I initially thought
I was doing something wrong with managing the <list.h> ring, so I that
code using old fashion double-link list.  The ring management wasn't the
problem, after all, but I prefer the old-fashion d-link implementation
as it's easier to read.

2004/01/10 16:44:01-05:00 scott.feldman
[netdrvr e100] netpoll + fixes to speed/duplex forced settings

* Preserve forced speed/duplex settings during close/open sequence.
* Reset PHY when going from autoneg to forced speed/duplex settings
  using ethtool.
* Add netpoll support.

2004/01/10 16:25:23-05:00 jgarzik
[netdrvr e100] include linux/moduleparam.h

Fixes build.

2004/01/10 16:19:08-05:00 scott.feldman
[e100] ICH6 IDs + ia64 memcpy fix + module_param

* Add ICH6 device IDs.  Devices funcionally equivalent to supported ICH5
  devices, but new IDs.
* Fixed unaligned access to casted skb->data (Matt Willcox
  [willy@debian.org]).
* MODULE_PARM -> module_param
* Bug printk after register_netdev to identify nic details.
* misc cleanups.

2004/01/10 15:56:28-05:00 scott.feldman
[netdrvr] e100 version 3 (complete rewrite)

BKrev: 40359d4bER3gLqLG4RBjo4heg6YKtQ