[PATCH] Fix the display of max-ever-used-stack in sysrq-T output
The sysrq-T output currently tries to display the mimimum amount of free
stack which each task has ever had available.  It has been busted for years,
because we forgot to zero out the stack when it is first created.

Fix that up, adding a conig option for it.

If the option is disabled, or the arch is not x86 then the free stack usage
will display as zero.

BKrev: 40359e21IN4OWYkXgltd2EWCVLFDkQ