[PATCH] initial support for transmeta's efficeon processors
This implements the initial support for the AGP gart in Transmeta's
Efficeon processors.

This code is based on linus' code from a while ago.  Peter and I adapted
it for the current kernel.

The note at the top indicates how it was tested, known issues, and so
forth.  It is working quite well, and we'll work to fix the minor issues
and test it some more with s3, more cards, other southbridges, etc.,
next.

BKrev: 4035a5e2G9dJ6-EP48AQ5I_pDAJXrg