[PATCH] add range checking to sys_sysctl()
If you pass a huge nlen value into sys_sysctl() it will loop for hours
printing billions of question marks, so range-check the arguments.

Also, remove the code which informs the user that the sysctl is obsolete: it
allows unprivileged users to spam the logs.

BKrev: 4039a617T9iBDJk20LRXKtnISAHQ-Q