[PATCH] tty utf8 mode
This sounds like a good moment to come with this patch.

This is essentially a five-year-old patch by Bruno Haible.  It
introduces utf8 mode, and the effect it has on erasing input characters.

(Side note - without stty support, this is not very useful, but now
the infrastructure is in place)

BKrev: 403a9c07qV9K8swscddPnLRJLI0IVw