Merge redhat.com:/spare/repo/netdev-2.6/pcnet32 into redhat.com:/spare/repo/net-drivers-2.5
2004/02/24 19:29:26-05:00 brazilnut
[PATCH] pcnet32.c fix compile error

Arrgh.  I sent the wrong diff file.  This one fixes the compile error I
introduced ;-(

2004/02/24 19:29:19-05:00 brazilnut
[PATCH] whitespace only change to pcnet32.c

Last change for a while.  This modifies the whitespace only.
consistent indentation, deletes trailing tabs, trailing spaces, etc.

2004/02/24 19:29:12-05:00 brazilnut
[PATCH] pcnet32.c adds loopback test

This adds the ability to do a local loopback test for the adapter.  This
source was submitted by James Lewis <jklewis@us.ibm.com>.  Tested on PPC
and IA32 systems.

2004/02/24 19:29:04-05:00 brazilnut
[PATCH] pcnet32.c add PCI hot remove support

This patch adds support for PCI hot remove.  Tested on PPC64 boxes.
The driver has also been checked to comply with PCMCIA/cardbus remove
requirements, but as I don't have a PCMCIA card, it is untested.
If a PCMCIA card is removed, reads will return FFFF.  The FFFF will
cause all while loops to exit.

2004/02/24 19:28:57-05:00 brazilnut
[PATCH] pcnet32.c non-mii errors with ethtool

pcnet32 devices which do not contain an mii respond incorrectly to
ethtool commands.

Reported by VANDROVE@vc.cvut.cz and Thomas Munck Steenholdt.

2004/02/24 19:28:50-05:00 brazilnut
[PATCH] pcnet32.c handle failures in open

If there are errors in the open routine, the driver does not correctly clean
up and free resources.

This also removes an unnecessary netif_wake_queue, correctly supports the
debug parameter, and updates the version and date.

2004/02/24 19:28:42-05:00 brazilnut
[PATCH] 2.6.3 pcnet32.c change to use ethtool_ops

This changes the driver to use ethtool_ops to be more modular.  There is a
net delta of just a one line between the two implementations, but permission
checking and copying to/from user space, etc. is all done in a consistent
manner by ethtool.c

2004/02/24 19:28:35-05:00 brazilnut
[PATCH] 2.6.3 pcnet32.c convert to use netif_msg_*

This patch converts the driver to use netif_msg_* for printing messages.
Tested IA32.

2004/02/24 19:28:26-05:00 brazilnut
[PATCH] 2.6.3 pcnet32.c wrong vendor ID fix

This patch was originally sent by Jon Mason on Sat Jan 31.

"I created a patch that allows pcnet32 adapters sold in IBM RS/6000 systems
to be recognized in non-ppc computers.  The problem is that the Vendor ID
of this adapter is incorrect.  There is a ppc specific workaround to get
it working on that architecture (checkout
http://lxr.linux.no/source/arch/ppc/kernel/pci.c?a=ppc#L87
for the workaround), but it obviously won't fix the problem in non-ppc
architectures."

BKrev: 403bf5b5dEfqhWsZX622ORGrxUG-XQ