[PATCH] Make insert_resource work for alder IOAPIC resources
From: James Bottomley <James.Bottomley@SteelEye.com>

This is a necessary precursor patch for getting the Intel Alder motherboard
working (it has a PCI device corresponding to the IO-APIC which has to be
forcibly inserted into the machine's reserved memory region).

Eric Biederman was going to come up with a more comprehensive fix, but in
the meantime, this is the minimum necessary to get insert_resource to work
when the covering region is larger than the resource being inserted.

BKrev: 403c1e7eeIaU-cDqV_9DA0gT4cK_fg