[PATCH] ppc64: Fix a sleeping with spinlock bug in ioremap
From: Benjamin Herrenschmidt <benh@kernel.crashing.org>

ppc64 uses its own range allocator for ioremap (in order to allocate
things in a different space than normal vmalloc).  This is historic stuff,
we may get rid of it, but in the meantime, here's a patch turning the
spinlock in there into a semaphore so it doesn't blow up when doing
kmallocs.

BKrev: 403ee7661-2iQfGG8-_KfHEY6KBMJg