Merge bk://gkernel.bkbits.net/libata-2.5 into ppc970.osdl.org:/home/torvalds/v2.5/linux
2004/02/26 22:36:03-05:00 jgarzik
[libata ata_piix] Update PCI quirk with new Intel SATA devices

And also make a note in ata_piix.c so this is not forgotten
in the future.

2004/02/26 21:45:39-05:00 jgarzik
[libata ata_piix] add support for ICH6

2004/02/26 21:26:13-05:00 jgarzik
[libata] limit S/G table size to 128 entries

Agreeing with Ben H's recommendation, I reduce the max-hw-segments
number to 128, to match max-phys-segments number.  This shouldn't
affect performance because 128*64K is far above the max transfer size
for most current IDE devices.  Even 128*4K is OK.

BKrev: 403eeaa3YxR86NKj7Ut6d-W0lYE4dg