Merge bk://bk.arm.linux.org.uk/linux-2.6-rmk into ppc970.osdl.org:/home/torvalds/v2.5/linux
2004/02/27 18:16:49+00:00 rmk
[ARM] Make free_memmap() use PFNs instead of physical addresses.

2004/02/27 17:18:54+00:00 rmk
[ARM] amba_{request,release}_regions

Add documentation for the above two functions.  Allow a NULL name to
be passed to amba_request_regions.  In this case, we use the driver's
name associated with the device.

2004/02/27 17:04:55+00:00 rmk
[ARM] Add resources and platform devices for SA11x0 serial ports.

2004/02/27 16:37:12+00:00 rmk
[ARM] Allow decompressor to use "cache type" register

This allows the decompressor to automatically detect the geometry of
the cache while booting to ensure correct behaviour over a wider
number of ARM CPUs.

2004/02/27 15:50:02+00:00 rmk
[ARM] Optimise readsl.

This optimises the underlying string-long read function for little
endian platforms.

2004/02/27 15:41:20+00:00 rmk
[ARM] Improve bad IRQ reporting.

This gives greater information on the cause of the bad IRQ, allowing
the cause to be more effectively traced.

2004/02/27 15:27:48+00:00 rmk
[ARM] Update mach-types definitions file.

2004/02/27 15:22:13+00:00 rmk
[ARM] Remove obsolete sysctl PM interface.

The old sysctl PM interface has been replaced by the driver models
interface.  This has been gone for a while in the -rmk tree and no
one complained, so there's no point keeping it around.

2004/02/27 15:16:18+00:00 rmk
[ARM] Add SA11x0 sched_clock() implementation.

2004/02/27 14:02:35+00:00 rmk
[ARM] Update sa1111-pcibuf for dmapool changes.

- use dev_dbg for device-centric debugging messages
- use pr_debug for general debugging messages
- use dmapools instead of pcipools
- use NULL rather than 0 for NULL pointers
- use enum dma_data_direction rather than int
- use DMA_* direction definitions rather than PCI_DMA_*
- only check for sane DMA direction on mapping functions, but
  check that DMA direction matches when unmapping/syncing.

2004/02/27 10:16:19+00:00 rmk
[ARM] Cleanup MODULE_* macros

Thanks to Adrian Bunk.

Remove kernel 2.0 #ifdef's from arm code, and move MODULE_* to
the end of the file.  Add an appropriate MODULE_LICENSE().

2004/02/26 14:24:43+00:00 rmk
[ARM] Add ARM architecture version 6 support.

This cset adds support ARM architecture version 6.

2004/02/26 12:15:46+00:00 armcc2000
[ARM] Optimise ARM720T Thumb abort unwinding.

Patch from: Andre.

Optimise data_thumb_pushpop and data_thumb_ldmstm population
counting code with a slightly more optimal algorithm than the
original as there are only 8 bits to count instead of 16.

BKrev: 403f9ab7N_9xFiYbiLoqV5kBkQ8lYQ