Merge bk://kernel.bkbits.net/gregkh/linux/pci-2.6 into ppc970.osdl.org:/home/torvalds/v2.5/linux
2004/02/24 14:04:40-08:00 greg
PCI Hotplug: clean up the Makefile a bit more.

Still need to fix up the pcie and shpc options to be saner for distros to use.

2004/02/24 14:02:08-08:00 greg
PCI Hotplug: remove unneeded ACPI Makefile rules.

2004/02/24 11:08:29-08:00 dlsy
[PATCH] PCI Hotplug: Patch to get polling mode in SHPC hot-plug driver properly working

Here is the patch to get polling mode in SHPC hot-plug properly
working.

2004/02/24 11:08:13-08:00 lxiep
[PATCH] PCI Hotplug: fix rpaphp bugs

The attached patch has a fix for the conflict between  fakephp and
rpaphp so that fakephp and rpaphp can't be built into the kernel at the
same time and a couple of  problems I missed in my previous patch.
(Sorry about that).

2004/02/24 11:07:59-08:00 rmk-pci
[PATCH] PCI: Introduce bus->bridge_ctl member

GregKH mentioned confirmed that people have been waiting for this patch.
Appologies, it had completely evaporated from my mind.

This patch introduces the "bridge_ctl" member of pci_bus, which allows
architectures to tweak the bridge control register (eg, for setting
fast back to back modes etc) in pcibios_fixup_bus().

Please note, though, that the value is only written back if
pci_setup_bridge() is called.  This will be called if an architecture
is using the generic PCI bus setup functionality in setup-bus.c.
If an architecture doesn't, then it is the responsibility of the
architecture to write this value to the bridge as appropriate.

(That said, the bridge control register is only ever changed if an
architecture is using setup-bus.c anyway, so there should be no
overall functional change.)

2004/02/24 11:07:43-08:00 rmk-pci
[PATCH] PCI: Don't report pci_request_regions() failure twice

pci_request_regions() reports an error when pci_request_region() fails.
However, since pci_request_region() already reports an error on failure,
pci_request_regions() has some unwanted duplication.

2004/02/24 11:07:29-08:00 rmk-pci
[PATCH] PCI: Report meaningful error for failed resource allocation

pci_assign_resource reports odd messages when resource allocation fails.
This is because res->end and res->start are modified by allocate_resource.
For example:

PCI: Failed to allocate resource 1(0-ffffffff) for 0000:00:01.0

The following patch reports whether it's an IO or memory resource, and
includes the correct size.  For consistency, we report it in a similar
way to the failure message in pci_request_region(), even though
res->start is unlikely to be useful.

BKrev: 403fed6d22Rz0htY4raHwpxI7zEhMg