[PATCH] fix in-place de/encryption bug with highmem
From: Christophe Saout <christophe@saout.de>

This patch fixes the bug where in-place encryption was not detected when
the same highmem pages is mapped twice to different virtual addresses.

This adds a parameter to xxx_process to indicate whether this is an
in-place encryption and moves the responsability to the caller using a
helper function scatterwalk.h.

BKrev: 40443186BowXpsBHgFbxClLNqkdCyw