[PATCH] radeonfb: some more PLL problems
I've had reports of flicker that appear with large (23") flat panels
and radeonfb. From experiments, it appears that forbiding the "odd"
PLL divider values fix it (like it fixes the blur problem on TMDS2).

There should not be anything special with TMDS1 and "odd" PLL values
though, so the problem may be subtly different (a bandwidth problem),
but until I have proper bandwidth calculation and access to this
monitor, the following patch is an acceptable workaround (Odd PLL values
aren't that useful anyway)

BKrev: 40467149OeRt0ybtBsvaTTRZ9CdMrg