[PATCH] /proc/cpuinfo fixes for G5
This adds a "timbase" entry in /proc/cpuinfo like p/iSeries that
provides the CPU timebase frequency.  It is using by a all sort of
performance analysis tools we are hacking in house.

It also remove a useless bit about the l2 cache that was copied over
from ppc32.

BKrev: 404673d93jP42gNH3-XhPj8pbEM4Sw