[PATCH] stop_machine_run: Move Bogolock Code Out of module.c
The "bogolock" code was introduced in module.c, as a way of freezing
the machine when we wanted to remove a module.  This patch moves it
out to stop_machine.c and stop_machine.h.

Since the code changes affinity and proirity, it's impolite to hijack
the current context, so we use a kthread.  This means we have to pass
the function rather than implement "stop_machine()" and
"restart_machine()".

BKrev: 404d1a0dHLG7mSYtz5LX1iDB1cs5EQ