Merge bk://gkernel.bkbits.net/libata-2.5 into ppc970.osdl.org:/home/torvalds/v2.5/linux
2004/03/08 22:33:25-05:00 jgarzik
[libata sata_promise] provide proper SCSI completion function

Promise driver uses a custom error handling function, so we need
the fix that was applied to the libata core:  the SCSI error
handling thread requires that we complete commands using a special
completion function, since the normal one doesn't work inside
the error handling thread.

2004/03/08 22:23:20-05:00 jgarzik
[libata ata_piix] Make sure annoying BIOSen don't disable our interrupts

2004/03/08 22:06:54-05:00 jgarzik
[libata] Include linux/suspend.h

2004/03/08 21:56:57-05:00 arjanv
[PATCH] sata vs suspend-to-ram

The following patch is a very crude one to at least not make sata block
suspend-to-ram entirely. Probably you want more powermanagement handling as
well... but it's a start

BKrev: 404d5559CTsXHyCdjaBrd39Yrle30w