[PATCH] Fix PCI<->OF matching on G5 AGP bus
Strangely, I though I fixed that a long time ago, but it was still
broken in the current tree...

Drivers like radeonfb fail to find the OF device matching a given PCI
device on the G5 AGP bus because of some bus renumbering tricks.  This
patch fixes the problem by fixing the bus numbers in the OF node.  This
corrects radeonfb and other drivers looking for EDID / PLL datas in the
OF node.

BKrev: 404e83a68P8L5Q8-x0lMoYwUBiHfHQ