[PATCH] ppc64: Let OF initialize all displays in the system
This patch reworks the early boot calls to OF to initialize displays.

All present displays are now initialized in reverse order so the
OF console stays on the first one. Initializing them all is necessary
for dual head configurations as we need OF driver to properly setup
the secondary TMDS of the video card, XFree isn't able to do that
currently

BKrev: 404e9638utiVdNq2y62EWQyi1f4kyw